+++ packages/webview_flutter/webview_flutter_android/test/android_webview_controller_test.mocks.dart
@@ -720,22 +720,26 @@ class MockAndroidWebViewProxy extends _i1.Mock
     void Function(
       _i2.WebView,
       int,
-    )? onProgressChanged,
+    )?
+        onProgressChanged,
     _i9.Future<List<String>> Function(
       _i2.WebView,
       _i2.FileChooserParams,
-    )? onShowFileChooser,
+    )?
+        onShowFileChooser,
   }) get createAndroidWebChromeClient => (super.noSuchMethod(
         Invocation.getter(#createAndroidWebChromeClient),
         returnValue: ({
           void Function(
             _i2.WebView,
             int,
-          )? onProgressChanged,
+          )?
+              onProgressChanged,
           _i9.Future<List<String>> Function(
             _i2.WebView,
             _i2.FileChooserParams,
-          )? onShowFileChooser,
+          )?
+              onShowFileChooser,
         }) =>
             _FakeWebChromeClient_0(
           this,
@@ -745,11 +749,13 @@ class MockAndroidWebViewProxy extends _i1.Mock
           void Function(
             _i2.WebView,
             int,
-          )? onProgressChanged,
+          )?
+              onProgressChanged,
           _i9.Future<List<String>> Function(
             _i2.WebView,
             _i2.FileChooserParams,
-          )? onShowFileChooser,
+          )?
+              onShowFileChooser,
         }) =>
             _FakeWebChromeClient_0(
           this,
@@ -759,28 +765,33 @@ class MockAndroidWebViewProxy extends _i1.Mock
         void Function(
           _i2.WebView,
           int,
-        )? onProgressChanged,
+        )?
+            onProgressChanged,
         _i9.Future<List<String>> Function(
           _i2.WebView,
           _i2.FileChooserParams,
-        )? onShowFileChooser,
+        )?
+            onShowFileChooser,
       }));
   @override
   _i2.WebViewClient Function({
     void Function(
       _i2.WebView,
       String,
-    )? onPageFinished,
+    )?
+        onPageFinished,
     void Function(
       _i2.WebView,
       String,
-    )? onPageStarted,
+    )?
+        onPageStarted,
     void Function(
       _i2.WebView,
       int,
       String,
       String,
-    )? onReceivedError,
+    )?
+        onReceivedError,
     void Function(
       _i2.WebView,
       _i2.WebResourceRequest,
@@ -791,32 +802,38 @@ class MockAndroidWebViewProxy extends _i1.Mock
       _i2.WebView,
       _i2.WebResourceRequest,
       _i2.WebResourceError,
-    )? onReceivedRequestError,
+    )?
+        onReceivedRequestError,
     void Function(
       _i2.WebView,
       _i2.WebResourceRequest,
-    )? requestLoading,
+    )?
+        requestLoading,
     void Function(
       _i2.WebView,
       String,
-    )? urlLoading,
+    )?
+        urlLoading,
   }) get createAndroidWebViewClient => (super.noSuchMethod(
         Invocation.getter(#createAndroidWebViewClient),
         returnValue: ({
           void Function(
             _i2.WebView,
             String,
-          )? onPageFinished,
+          )?
+              onPageFinished,
           void Function(
             _i2.WebView,
             String,
-          )? onPageStarted,
+          )?
+              onPageStarted,
           void Function(
             _i2.WebView,
             int,
             String,
             String,
-          )? onReceivedError,
+          )?
+              onReceivedError,
           void Function(
             _i2.WebView,
             _i2.WebResourceRequest,
@@ -827,15 +844,18 @@ class MockAndroidWebViewProxy extends _i1.Mock
             _i2.WebView,
             _i2.WebResourceRequest,
             _i2.WebResourceError,
-          )? onReceivedRequestError,
+          )?
+              onReceivedRequestError,
           void Function(
             _i2.WebView,
             _i2.WebResourceRequest,
-          )? requestLoading,
+          )?
+              requestLoading,
           void Function(
             _i2.WebView,
             String,
-          )? urlLoading,
+          )?
+              urlLoading,
         }) =>
             _FakeWebViewClient_1(
           this,
@@ -845,17 +865,20 @@ class MockAndroidWebViewProxy extends _i1.Mock
           void Function(
             _i2.WebView,
             String,
-          )? onPageFinished,
+          )?
+              onPageFinished,
           void Function(
             _i2.WebView,
             String,
-          )? onPageStarted,
+          )?
+              onPageStarted,
           void Function(
             _i2.WebView,
             int,
             String,
             String,
-          )? onReceivedError,
+          )?
+              onReceivedError,
           void Function(
             _i2.WebView,
             _i2.WebResourceRequest,
@@ -866,15 +889,18 @@ class MockAndroidWebViewProxy extends _i1.Mock
             _i2.WebView,
             _i2.WebResourceRequest,
             _i2.WebResourceError,
-          )? onReceivedRequestError,
+          )?
+              onReceivedRequestError,
           void Function(
             _i2.WebView,
             _i2.WebResourceRequest,
-          )? requestLoading,
+          )?
+              requestLoading,
           void Function(
             _i2.WebView,
             String,
-          )? urlLoading,
+          )?
+              urlLoading,
         }) =>
             _FakeWebViewClient_1(
           this,
@@ -884,17 +910,20 @@ class MockAndroidWebViewProxy extends _i1.Mock
         void Function(
           _i2.WebView,
           String,
-        )? onPageFinished,
+        )?
+            onPageFinished,
         void Function(
           _i2.WebView,
           String,
-        )? onPageStarted,
+        )?
+            onPageStarted,
         void Function(
           _i2.WebView,
           int,
           String,
           String,
-        )? onReceivedError,
+        )?
+            onReceivedError,
         void Function(
           _i2.WebView,
           _i2.WebResourceRequest,
@@ -905,15 +934,18 @@ class MockAndroidWebViewProxy extends _i1.Mock
           _i2.WebView,
           _i2.WebResourceRequest,
           _i2.WebResourceError,
-        )? onReceivedRequestError,
+        )?
+            onReceivedRequestError,
         void Function(
           _i2.WebView,
           _i2.WebResourceRequest,
-        )? requestLoading,
+        )?
+            requestLoading,
         void Function(
           _i2.WebView,
           String,
-        )? urlLoading,
+        )?
+            urlLoading,
       }));
   @override
   _i2.FlutterAssetManager Function() get createFlutterAssetManager =>
@@ -962,7 +994,8 @@ class MockAndroidWebViewProxy extends _i1.Mock
         String,
         String,
         int,
-      ) onDownloadStart}) get createDownloadListener => (super.noSuchMethod(
+      )
+          onDownloadStart}) get createDownloadListener => (super.noSuchMethod(
         Invocation.getter(#createDownloadListener),
         returnValue: (
                 {required void Function(
@@ -971,7 +1004,8 @@ class MockAndroidWebViewProxy extends _i1.Mock
                   String,
                   String,
                   int,
-                ) onDownloadStart}) =>
+                )
+                    onDownloadStart}) =>
             _FakeDownloadListener_2(
           this,
           Invocation.getter(#createDownloadListener),
@@ -983,7 +1017,8 @@ class MockAndroidWebViewProxy extends _i1.Mock
                   String,
                   String,
                   int,
-                ) onDownloadStart}) =>
+                )
+                    onDownloadStart}) =>
             _FakeDownloadListener_2(
           this,
           Invocation.getter(#createDownloadListener),
@@ -995,7 +1030,8 @@ class MockAndroidWebViewProxy extends _i1.Mock
             String,
             String,
             int,
-          ) onDownloadStart}));
+          )
+              onDownloadStart}));
   @override
   _i9.Future<void> setWebContentsDebuggingEnabled(bool? enabled) =>
       (super.noSuchMethod(
@@ -1068,8 +1104,8 @@ class MockAndroidWebViewWidgetCreationParams extends _i1.Mock
       (super.noSuchMethod(
         Invocation.getter(#gestureRecognizers),
         returnValue: <_i11.Factory<_i12.OneSequenceGestureRecognizer>>{},
-        returnValueForMissingStub: <_i11
-            .Factory<_i12.OneSequenceGestureRecognizer>>{},
+        returnValueForMissingStub: <
+            _i11.Factory<_i12.OneSequenceGestureRecognizer>>{},
       ) as Set<_i11.Factory<_i12.OneSequenceGestureRecognizer>>);
 }