+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/legacy/web_kit_webview_widget.dart
@@ -655,7 +655,8 @@ class WebViewWidgetProxy {
       String keyPath,
       NSObject object,
       Map<NSKeyValueChangeKey, Object?> change,
-    )? observeValue,
+    )?
+        observeValue,
   }) {
     return WKWebView(configuration, observeValue: observeValue);
   }
@@ -665,7 +666,8 @@ class WebViewWidgetProxy {
     required void Function(
       WKUserContentController userContentController,
       WKScriptMessage message,
-    ) didReceiveScriptMessage,
+    )
+        didReceiveScriptMessage,
   }) {
     return WKScriptMessageHandler(
       didReceiveScriptMessage: didReceiveScriptMessage,
@@ -678,7 +680,8 @@ class WebViewWidgetProxy {
       WKWebView webView,
       WKWebViewConfiguration configuration,
       WKNavigationAction navigationAction,
-    )? onCreateWebView,
+    )?
+        onCreateWebView,
   }) {
     return WKUIDelegate(onCreateWebView: onCreateWebView);
   }
@@ -691,7 +694,8 @@ class WebViewWidgetProxy {
     Future<WKNavigationActionPolicy> Function(
       WKWebView webView,
       WKNavigationAction navigationAction,
-    )? decidePolicyForNavigationAction,
+    )?
+        decidePolicyForNavigationAction,
     void Function(WKWebView webView, NSError error)? didFailNavigation,
     void Function(WKWebView webView, NSError error)?
         didFailProvisionalNavigation,