+++ packages/webview_flutter/webview_flutter_wkwebview/test/legacy/web_kit_webview_widget_test.dart
@@ -221,7 +221,8 @@ void main() {
         );
 
         verify(mockWebViewConfiguration
-            .setMediaTypesRequiringUserActionForPlayback(<WKAudiovisualMediaType>{
+            .setMediaTypesRequiringUserActionForPlayback(<
+                WKAudiovisualMediaType>{
           WKAudiovisualMediaType.all,
         }));
       });
@@ -239,7 +240,8 @@ void main() {
         );
 
         verify(mockWebViewConfiguration
-            .setMediaTypesRequiringUserActionForPlayback(<WKAudiovisualMediaType>{
+            .setMediaTypesRequiringUserActionForPlayback(<
+                WKAudiovisualMediaType>{
           WKAudiovisualMediaType.none,
         }));
       });