+++ packages/webview_flutter/webview_flutter_wkwebview/CHANGELOG.md
@@ -3,12 +3,12 @@
 * Updates pigeon to fix warnings with clang 15.
 * Updates minimum Flutter version to 3.3.
 * Fixes common typos in tests and documentation.
+* Adds `decidePolicyForNavigationResponse` to internal WKNavigationDelegate to support the
+  `onHttpError` callback.
 
 ## 3.2.3
 
 * Updates to `pigeon` version 7.
-* Adds `decidePolicyForNavigationResponse` to internal WKNavigationDelegate to support the
-  `onHttpError` callback.
 
 ## 3.2.2